+++ graph/ops_convolve_test.go
@@ -20,7 +20,7 @@ import (
 	"testing"
 
 	. "github.com/gomlx/gomlx/graph"
-	"github.com/gomlx/gomlx/types/tensors"
+	"github.com/gomlx/gomlx/types/shapes"
 	"github.com/gomlx/gomlx/types/tensors/images"
 	"github.com/gomlx/gopjrt/dtypes"
 )
@@ -110,7 +110,7 @@ func TestConvolve(t *testing.T) {
 }
 
 func TestGradientConvolve(t *testing.T) {
-	testGradients(t, "Gradient 1D Convolve().NoPadding() -- scaled output",
+	testGradients(t, "1D.NoPadding()",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := Add(IotaFull(g, MakeShape(dtypes.Float64, 1, 5, 1)), Const(g, 1.0))
 			kernel := Add(IotaFull(g, MakeShape(dtypes.Float64, 3, 1, 1)), Const(g, 1.0))
@@ -124,7 +124,7 @@ func TestGradientConvolve(t *testing.T) {
 			[][][]float64{{{14}}, {{20}}, {{26}}},
 		})
 
-	testGradients(t, "Gradient 2D Convolve().NoPadding()",
+	testGradients(t, "2D.NoPadding()",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			channelA := IotaFull(g, MakeShape(dtypes.Float64, 1, 5, 5, 1))
 			channelB := Mul(channelA, Const(g, 0.001))
@@ -149,7 +149,7 @@ func TestGradientConvolve(t *testing.T) {
 			},
 		})
 
-	testGradients(t, "Gradient Convolve().PadSame()",
+	testGradients(t, "2D.PadSame()",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			channelA := IotaFull(g, MakeShape(dtypes.Float64, 1, 3, 3, 1))
 			channelB := Mul(channelA, Const(g, 0.1))
@@ -171,7 +171,7 @@ func TestGradientConvolve(t *testing.T) {
 			},
 		})
 
-	testGradients(t, "Gradient Convolve().NoPadding().Dilations(2)",
+	testGradients(t, "2D.NoPadding().Dilations(2)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			channelA := IotaFull(g, MakeShape(dtypes.Float64, 1, 5, 5, 1))
 			channelB := Mul(channelA, Const(g, 0.001))
@@ -195,7 +195,7 @@ func TestGradientConvolve(t *testing.T) {
 			},
 		})
 
-	testGradients(t, "Gradient 2D Convolve().NoPadding().Strides(2)",
+	testGradients(t, "2D.NoPadding().Strides(2)",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			channelA := IotaFull(g, MakeShape(dtypes.Float64, 1, 5, 5, 1))
 			channelB := Mul(channelA, Const(g, 0.001))
@@ -219,26 +219,26 @@ func TestGradientConvolve(t *testing.T) {
 			},
 		})
 
-	testGradients(t, "Gradient 2D Convolve().NoPadding().Strides(2): shape check",
+	testGradients(t, "2D.NoPadding().Strides(2): shape check",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := Zeros(g, MakeShape(dtypes.Float64, 2, 8, 7, 3))
 			kernel := Zeros(g, MakeShape(dtypes.Float64, 1, 1, 3, 6))
 			output = Convolve(input, kernel).NoPadding().Strides(2).Done()
 			return output, []*Node{input, kernel}
 		}, []any{
-			tensors.FromScalarAndDimensions(0.0, 2, 8, 7, 3).Value(),
-			tensors.FromScalarAndDimensions(0.0, 1, 1, 3, 6).Value(),
+			shapes.Make(dtypes.Float64, 2, 8, 7, 3),
+			shapes.Make(dtypes.Float64, 1, 1, 3, 6),
 		})
 
-	testGradients(t, "Gradient 2D Convolve().NoPadding().Dilations(2): shape check",
+	testGradients(t, "2D.NoPadding().Dilations(2): shape check",
 		func(g *Graph) (output *Node, nodesForGrad []*Node) {
 			input := Zeros(g, MakeShape(dtypes.Float64, 2, 5, 5, 3))
 			kernel := Zeros(g, MakeShape(dtypes.Float64, 2, 2, 3, 6))
 			output = Convolve(input, kernel).NoPadding().Dilations(2).Done()
 			return output, []*Node{input, kernel}
 		}, []any{
-			tensors.FromScalarAndDimensions(0.0, 2, 5, 5, 3).Value(),
-			tensors.FromScalarAndDimensions(0.0, 2, 2, 3, 6).Value(),
+			shapes.Make(dtypes.Float64, 2, 5, 5, 3),
+			shapes.Make(dtypes.Float64, 2, 2, 3, 6),
 		})
 }
 
@@ -277,4 +277,90 @@ func TestConvolveWithGrouping(t *testing.T) {
 			return
 		},
 		[][][][]float32{{{{4.0}}, {{26.0}}}})
+
+	testGradients(t, "Gradient-Input:ChannelGroupCount(2)",
+		func(g *Graph) (output *Node, nodesForGrad []*Node) {
+			input := Ones(g, MakeShape(F32, 1, 5, 2))
+			kernel := IotaFull(g, MakeShape(F32, 3, 1, 4))
+			convNode := Convolve(input, kernel).
+				ChannelGroupCount(2).
+				Strides(1).
+				NoPadding().
+				Done()
+			output = ReduceAllSum(convNode)
+			return output, []*Node{input}
+		}, []any{
+			[][][]float32{{
+				{1., 5.},
+				{10., 18.},
+				{27., 39.},
+				{26., 34.},
+				{17., 21.},
+			}},
+		})
+
+	testGradients(t, "Gradient-Kernel:ChannelGroupCount(2)",
+		func(g *Graph) (output *Node, nodesForGrad []*Node) {
+			input := IotaFull(g, MakeShape(F32, 1, 5, 2))
+			kernel := Ones(g, MakeShape(F32, 3, 1, 4))
+			convNode := Convolve(input, kernel).
+				ChannelGroupCount(2).
+				Strides(1).
+				NoPadding().
+				Done()
+			output = ReduceAllSum(convNode)
+			return output, []*Node{kernel}
+		}, []any{
+			[][][]float32{
+				{{6., 6., 9., 9.}},
+				{{12., 12., 15., 15.}},
+				{{18., 18., 21., 21.}},
+			},
+		})
+
+	testGradients(t, "Gradient-Input:BatchGroupCount(2)",
+		func(g *Graph) (output *Node, nodesForGrad []*Node) {
+			input := Ones(g, MakeShape(F32, 2, 5, 1))
+			kernel := IotaFull(g, MakeShape(F32, 3, 1, 4))
+			convNode := Convolve(input, kernel).
+				BatchGroupCount(2).
+				Strides(1).
+				NoPadding().
+				Done()
+			output = ReduceAllSum(convNode)
+			return output, []*Node{input}
+		}, []any{
+			[][][]float32{{
+				{1.},
+				{10.},
+				{27.},
+				{26.},
+				{17.},
+			}, {
+				{5.},
+				{18.},
+				{39.},
+				{34.},
+				{21.},
+			}},
+		})
+
+	testGradients(t, "Gradient-Kernel:BatchGroupCount(2)",
+		func(g *Graph) (output *Node, nodesForGrad []*Node) {
+			input := IotaFull(g, MakeShape(F32, 2, 5, 1))
+			kernel := Ones(g, MakeShape(F32, 3, 1, 4))
+			convNode := Convolve(input, kernel).
+				BatchGroupCount(2).
+				Strides(1).
+				NoPadding().
+				Done()
+			output = ReduceAllSum(convNode)
+			return output, []*Node{kernel}
+		}, []any{
+			[][][]float32{
+				{{3., 3., 18., 18.}},
+				{{6., 6., 21., 21.}},
+				{{9., 9., 24., 24.}},
+			},
+		})
 }